create new private file _structured_events.dart
