+++ packages/devtools_extensions/lib/src/api/_structured_events.dart
@@ -1,0 +1,89 @@
+// Copyright 2023 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'api.dart';
+import 'model.dart';
+
+/// An extension event of type [DevToolsExtensionEventType.showNotification]
+/// that is sent from an extension to DevTools asking DevTools to post a
+/// notification the the DevTools notification framework.
+class ShowNotificationExtensionEvent extends DevToolsExtensionEvent {
+  ShowNotificationExtensionEvent({required String message})
+      : super(
+          DevToolsExtensionEventType.showNotification,
+          data: {_messageKey: message},
+        );
+
+  factory ShowNotificationExtensionEvent.from(DevToolsExtensionEvent event) {
+    assert(event.type == DevToolsExtensionEventType.showNotification);
+    final message = event.data!.checkValid<String>(_messageKey);
+    return ShowNotificationExtensionEvent(message: message);
+  }
+
+  static const _messageKey = 'message';
+
+  String get message => data![_messageKey] as String;
+}
+
+/// An extension event of type [DevToolsExtensionEventType.showBannerMessage]
+/// that is sent from an extension to DevTools asking DevTools to post a
+/// banner message to the extension's screen using the DevTools banner message
+/// framework.
+class ShowBannerMessageExtensionEvent extends DevToolsExtensionEvent {
+  ShowBannerMessageExtensionEvent({
+    required String id,
+    required String bannerMessageType,
+    required String message,
+    required String extensionName,
+  }) : super(
+          DevToolsExtensionEventType.showBannerMessage,
+          data: {
+            _idKey: id,
+            _bannerMessageTypeKey: bannerMessageType,
+            _messageKey: message,
+            _extensionNameKey: extensionName,
+          },
+        );
+
+  factory ShowBannerMessageExtensionEvent.from(DevToolsExtensionEvent event) {
+    assert(event.type == DevToolsExtensionEventType.showBannerMessage);
+    final eventData = event.data!;
+    final id = eventData.checkValid<String>(_idKey);
+    final message = eventData.checkValid<String>(_messageKey);
+    final type = eventData.checkValid<String>(_bannerMessageTypeKey);
+    final extensionName = eventData.checkValid<String>(_extensionNameKey);
+    return ShowBannerMessageExtensionEvent(
+      id: id,
+      bannerMessageType: type,
+      message: message,
+      extensionName: extensionName,
+    );
+  }
+
+  static const _messageKey = 'message';
+  static const _idKey = 'id';
+  static const _bannerMessageTypeKey = 'bannerMessageType';
+  static const _extensionNameKey = 'extensionName';
+
+  String get messageId => data![_idKey] as String;
+  String get bannerMessageType => data![_bannerMessageTypeKey] as String;
+  String get message => data![_messageKey] as String;
+  String get extensionName => data![_extensionNameKey] as String;
+}
+
+extension ParseExtension on Map<String, Object?> {
+  T checkValid<T>(String key) {
+    final element = this[key];
+    if (element == null) {
+      throw FormatException("Missing key '$key'");
+    }
+    if (element is! T) {
+      throw FormatException(
+        'Expected element of type $T but got element of type '
+        '${element.runtimeType}.',
+      );
+    }
+    return element as T;
+  }
+}
