+++ packages/devtools_extensions/lib/src/api/model.dart
@@ -54,78 +54,3 @@ class DevToolsExtensionEvent {
 /// A void callback that handles a [DevToolsExtensionEvent].
 typedef ExtensionEventHandler = void Function(DevToolsExtensionEvent event);
 
-class ShowNotificationExtensionEvent extends DevToolsExtensionEvent {
-  ShowNotificationExtensionEvent({required String message})
-      : super(
-          DevToolsExtensionEventType.showNotification,
-          data: {_messageKey: message},
-        );
-
-  factory ShowNotificationExtensionEvent.from(DevToolsExtensionEvent event) {
-    assert(event.type == DevToolsExtensionEventType.showNotification);
-    final message = event.data!.checkValid<String>(_messageKey);
-    return ShowNotificationExtensionEvent(message: message);
-  }
-
-  static const _messageKey = 'message';
-
-  String get message => data![_messageKey] as String;
-}
-
-class ShowBannerMessageExtensionEvent extends DevToolsExtensionEvent {
-  ShowBannerMessageExtensionEvent({
-    required String id,
-    required String bannerMessageType,
-    required String message,
-    required String extensionName,
-  }) : super(
-          DevToolsExtensionEventType.showBannerMessage,
-          data: {
-            _idKey: id,
-            _bannerMessageTypeKey: bannerMessageType,
-            _messageKey: message,
-            _extensionNameKey: extensionName,
-          },
-        );
-
-  factory ShowBannerMessageExtensionEvent.from(DevToolsExtensionEvent event) {
-    assert(event.type == DevToolsExtensionEventType.showBannerMessage);
-    final eventData = event.data!;
-    final id = eventData.checkValid<String>(_idKey);
-    final message = eventData.checkValid<String>(_messageKey);
-    final type = eventData.checkValid<String>(_bannerMessageTypeKey);
-    final extensionName = eventData.checkValid<String>(_extensionNameKey);
-    return ShowBannerMessageExtensionEvent(
-      id: id,
-      bannerMessageType: type,
-      message: message,
-      extensionName: extensionName,
-    );
-  }
-
-  static const _messageKey = 'message';
-  static const _idKey = 'id';
-  static const _bannerMessageTypeKey = 'bannerMessageType';
-  static const _extensionNameKey = 'extensionName';
-
-  String get messageId => data![_idKey] as String;
-  String get bannerMessageType => data![_bannerMessageTypeKey] as String;
-  String get message => data![_messageKey] as String;
-  String get extensionName => data![_extensionNameKey] as String;
-}
-
-extension ParseExtension on Map<String, Object?> {
-  T checkValid<T>(String key) {
-    final element = this[key];
-    if (element == null) {
-      throw FormatException("Missing key '$key'");
-    }
-    if (element is! T) {
-      throw FormatException(
-        'Expected element of type $T but got element of type '
-        '${element.runtimeType}.',
-      );
-    }
-    return element as T;
-  }
-}
